requirements.txt
pandas
stanza
emoji
nltk
deap
twisted
gensim
tweepy==4.3.0
python-dotenv
geopy
